Tl alex branch

+++ src/main/java/by/grsu/liceum/service/impl/AccountServiceImpl.java
@@ -7,17 +7,21 @@ import by.grsu.liceum.dto.account.AccountShortcutDto;
 import by.grsu.liceum.dto.mapper.AccountDtoMapper;
 import by.grsu.liceum.entity.Account;
 import by.grsu.liceum.entity.Card;
+import by.grsu.liceum.entity.Institution;
 import by.grsu.liceum.entity.Role;
-import by.grsu.liceum.entity.enums.RoleConstant;
 import by.grsu.liceum.exception.AccountWithIdNotFoundException;
+import by.grsu.liceum.exception.InstitutionWithIdNotFoundException;
+import by.grsu.liceum.exception.InvalidPermissionsException;
 import by.grsu.liceum.exception.InvalidRoleNameException;
 import by.grsu.liceum.repository.AccountRepository;
+import by.grsu.liceum.repository.InstitutionRepository;
 import by.grsu.liceum.repository.RoleRepository;
 import by.grsu.liceum.service.AccountService;
 import by.grsu.liceum.service.CardService;
 import by.grsu.liceum.utils.Generator;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,34 +33,45 @@ public class AccountServiceImpl implements AccountService {
     private final AccountRepository accountRepository;
     private final RoleRepository roleRepository;
     private final CardService cardService;
+    private final InstitutionRepository institutionRepository;
+    private final BCryptPasswordEncoder bCryptPasswordEncoder;
 
     @Override
-    public AccountFullDto findById(long id) {
+    public AccountFullDto findById(long institutionId, long id) {
         Account account = Optional.ofNullable(accountRepository.findById(id))
                 .orElseThrow(() -> new AccountWithIdNotFoundException(id));
+
+        if(account.getInstitution().getId() != institutionId)
+            throw new InvalidPermissionsException();
 
         return AccountDtoMapper.convertEntityToFullDto(account);
     }
 
     @Override
-    public List<AccountShortcutDto> findAll() {
-        return accountRepository.findAll().stream()
+    public List<AccountShortcutDto> findAll(long institutionId) {
+        return accountRepository.findAllByInstitution_Id(institutionId).stream()
                 .map(AccountDtoMapper::convertEntityToShortcutDto)
                 .toList();
     }
 
     @Override
-    @Transactional //TODO ADMIN_ROLE
-    public AccountCreationResponse createUserWithRole(AccountCreationDto creationDto) {
+    @Transactional
+    public AccountCreationResponse createUserWithRole(long institutionId, AccountCreationDto creationDto) {
+        Institution institution = Optional.ofNullable(institutionRepository.findById(institutionId))
+                .orElseThrow(() -> new InstitutionWithIdNotFoundException(institutionId));
+
         Account account = AccountDtoMapper.convertDtoToEntity(creationDto);
 
+        String password = Generator.generatePassword("ROLE_USER");
+
         account.setLogin(Generator.generateLogin(AccountDtoMapper.convertCreationDtoToGeneratorDto(creationDto)));
-        account.setPassword(Generator.generatePassword());
+        account.setPassword(bCryptPasswordEncoder.encode(password));
+        account.setInstitution(institution);
 
         accountRepository.save(account);
 
-        for(RoleConstant roleConstant : creationDto.getRoleNames()){
-            Role role = Optional.ofNullable(roleRepository.findByName(roleConstant))
+        for(String roleConstant : creationDto.getRoleNames()){
+            Role role = Optional.ofNullable(roleRepository.findByName(roleConstant.toUpperCase()))
                     .orElseThrow(() -> new InvalidRoleNameException(roleConstant));
 
             role.getAccounts().add(account);
@@ -67,13 +82,40 @@ public class AccountServiceImpl implements AccountService {
 
         account.setCard(card);
 
-        return AccountDtoMapper.convertEntityToCreationResponse(account);
+        institution.getAccounts().add(account);
+
+        AccountCreationResponse response = AccountDtoMapper.convertEntityToCreationResponse(account);
+        response.setPassword(password);
+
+        return response;
     }
 
     @Override
-    public void deleteById(long id) {
-        Optional.ofNullable(accountRepository.findById(id))
+    @Transactional
+    public AccountCreationResponse regeneratePassword(long institutionId, long accountId) {
+        Account account = Optional.ofNullable(accountRepository.findById(accountId))
+                .orElseThrow(() -> new AccountWithIdNotFoundException(accountId));
+
+        if(account.getInstitution().getId() != institutionId)
+            throw new InvalidPermissionsException();
+
+        String password = Generator.generatePassword("ROLE_USER");
+
+        account.setPassword(bCryptPasswordEncoder.encode(password));
+
+        AccountCreationResponse response = AccountDtoMapper.convertEntityToCreationResponse(account);
+        response.setPassword(password);
+
+        return response;
+    }
+
+    @Override
+    public void deleteById(long institutionId, long id) {
+        Account account = Optional.ofNullable(accountRepository.findById(id))
                 .orElseThrow(() -> new AccountWithIdNotFoundException(id));
+
+        if(account.getInstitution().getId() != institutionId)
+            throw new InvalidPermissionsException();
 
         accountRepository.deleteById(id);
     }
